#160 Add package in modules
