+++ modules/security/src/main/java/com/gg_pigs/security/GPAuthenticationFilter.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.security.core.context.SecurityContextHolder;
