+++ modules/security/src/main/java/com/gg_pigs/security/GPAuthenticationProvider.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 import org.springframework.security.authentication.AuthenticationProvider;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
