+++ modules/security/src/main/java/com/gg_pigs/security/GPAuthenticationToken.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 import org.springframework.security.authentication.AbstractAuthenticationToken;
 import org.springframework.security.core.GrantedAuthority;
