+++ modules/security/src/main/java/com/gg_pigs/security/GPUnauthenticatedException.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 class GPUnauthenticatedException extends GPSecurityException {
     public GPUnauthenticatedException() {
         super();
