+++ modules/security/src/main/java/com/gg_pigs/security/GPUserDetails.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 import org.springframework.security.core.userdetails.UserDetails;
 
 public interface GPUserDetails extends UserDetails {
