+++ modules/security/src/main/java/com/gg_pigs/security/GPUserDetailsService.java
@@ -1,3 +1,5 @@
+package com.gg_pigs.security;
+
 import org.springframework.security.core.userdetails.UserDetailsService;
 
 import javax.servlet.http.HttpServletRequest;
